Merge with V1

See merge request kinnarastudio/kecak-workflow!3

+++ wflow-core/src/main/java/org/joget/apps/form/dao/FormDataDaoImpl.java
@@ -246,9 +246,7 @@ public class FormDataDaoImpl extends HibernateDaoSupport implements FormDataDao 
 
         try {
             String query = "SELECT e FROM " + tableName + " e ";
-            if (condition != null) {
-                query += condition;
-            }
+            query += ((condition != null && condition != "")? condition + " AND ":" WHERE ") + (FormUtil.PROPERTY_DELETED + " = false OR " + FormUtil.PROPERTY_DELETED + " is null") ;
 
             if ((sort != null && !sort.trim().isEmpty()) && !query.toLowerCase().contains("order by")) {
                 String sortProperty = sort;
@@ -341,7 +339,9 @@ public class FormDataDaoImpl extends HibernateDaoSupport implements FormDataDao 
         // get hibernate template
         Session session = getHibernateSession(tableName, tableName, null, ACTION_TYPE_LOAD);
         try {
-            Query q = session.createQuery("SELECT COUNT(*) FROM " + tableName + " e " + condition);
+            String query = "SELECT COUNT(*) FROM " + tableName + " e ";
+            query += ((condition != null && condition != "")? condition + " AND ":" WHERE ") + (FormUtil.PROPERTY_DELETED + " = false OR " + FormUtil.PROPERTY_DELETED + " is null");
+            Query q = session.createQuery(query);
 
             if (params != null) {
                 int i = 0;
@@ -451,6 +451,7 @@ public class FormDataDaoImpl extends HibernateDaoSupport implements FormDataDao 
         try {
             // save the form data
             for (FormRow row : rowSet) {
+                if(row.getDeleted() == null) row.setDeleted(false);
                 session.saveOrUpdate(entityName, row);
             }
             session.flush();
@@ -486,16 +487,41 @@ public class FormDataDaoImpl extends HibernateDaoSupport implements FormDataDao 
         closeSession(session);
     }
 
+    private void deleteByPK(String entityName, String tableName, String[] primaryKeyValues){
+        for (String pk : primaryKeyValues){
+            FormRowSet rowSet = internalFind(entityName,tableName,"WHERE id=?", new Object[] {pk},null,false,0,0);
+            if(!rowSet.isEmpty()){
+                FormRow row = rowSet.get(0);
+                row.setDeleted(true);
+                rowSet.add(row);
+                internalSaveOrUpdate(entityName,tableName,rowSet);
+            }
+        }
+    }
+
     /**
      * Delete form data by primary keys
      * @param form
      * @param primaryKeyValues 
      */
-    public void delete(Form form, String[] primaryKeyValues) {
+    public void delete(Form form, String[] primaryKeyValues){
+        delete(form,primaryKeyValues,false);
+    }
+
+    /**
+     *
+     * @param form
+     * @param primaryKeyValues
+     * @param isHardDelete true for deleting data from database, false to flag data as deleted
+     */
+    public void delete(Form form, String[] primaryKeyValues, Boolean isHardDelete) {
         String entityName = getFormEntityName(form);
         String tableName = getFormTableName(form);
-
-        internalDelete(entityName, tableName, primaryKeyValues);
+        if(isHardDelete){
+            internalDelete(entityName, tableName, primaryKeyValues);
+        } else {
+            deleteByPK(entityName, tableName, primaryKeyValues);
+        }
     }
     
     /**
@@ -504,33 +530,58 @@ public class FormDataDaoImpl extends HibernateDaoSupport implements FormDataDao 
      * @param tableName
      * @param primaryKeyValues 
      */
-    public void delete(String formDefId, String tableName, String[] primaryKeyValues) {
+    public void delete(String formDefId, String tableName, String[] primaryKeyValues){
+        delete(formDefId,tableName,primaryKeyValues,false);
+    }
+
+    /**
+     *
+     * @param formDefId
+     * @param tableName
+     * @param primaryKeyValues
+     * @param isHardDelete true for deleting data from database, false to flag data as deleted
+     */
+    public void delete(String formDefId, String tableName, String[] primaryKeyValues, Boolean isHardDelete) {
         String entityName = getFormEntityName(formDefId);
         String newTableName = getFormTableName(formDefId, tableName);
+        if(isHardDelete){
+            internalDelete(entityName, newTableName, primaryKeyValues);
+        } else {
+            deleteByPK(entityName, newTableName, primaryKeyValues);
+        }
 
-        internalDelete(entityName, newTableName, primaryKeyValues);
     }
-    
+
     /**
      * Delete form data by primary keys
      * @param formDefId
      * @param tableName
      * @param rows
      */
-    public void delete(String formDefId, String tableName, FormRowSet rows) {
+    public void delete(String formDefId, String tableName, FormRowSet rows){
+        delete(formDefId,tableName,rows,false);
+    }
+
+    public void delete(String formDefId, String tableName, FormRowSet rows, Boolean isHardDelete) {
         String entityName = getFormEntityName(formDefId);
         String newTableName = getFormTableName(formDefId, tableName);
-
-        // get hibernate template
-        Session session = getHibernateSession(entityName, newTableName, null, ACTION_TYPE_STORE);
-        try {
-            // save the form data
-            for (FormRow row : rows) {
-                session.delete(entityName, row);
+        if(isHardDelete){
+            // get hibernate template
+            Session session = getHibernateSession(entityName, newTableName, null, ACTION_TYPE_STORE);
+            try {
+                // save the form data
+                for (FormRow row : rows) {
+                    session.delete(entityName, row);
+                }
+                session.flush();
+            } finally {
+                closeSession(session);
             }
-            session.flush();
-        } finally {
-            closeSession(session);
+        } else {
+            for (FormRow row : rows) {
+                row.setDeleted(true);
+            }
+            internalSaveOrUpdate(entityName, newTableName, rows);
         }
     }
